addd

+++ src/main/java/com/driver/services/BlogService.java
@@ -29,7 +29,6 @@ public class BlogService {
         //find all blogs
          List<Blog> blogList = blogRepository1.findAll();
          return blogList;
-
     }
 
     public void createAndReturnBlog(Integer userId, String title, String content) {
@@ -39,19 +38,24 @@ public class BlogService {
 
         //Updating the userInformation and changing its blogs
 
+        //create a blog at the current time
         Blog blog = new Blog();
-        blog.setUser(userRepository1.findById(userId).get());
+        User user = userRepository1.findById(userId).get();
+
+        //updating the blog details
+        blog.setUser(user);
         blog.setTitle(title);
         blog.setContent(content);
         blog.setPubDate(new Date());
-        User user = userRepository1.findById(userId).get();
-        List<Blog> blogList = user.getBlogList();
-        blogList.add(blog);
-        user.setBlogList(blogList);
-        blogRepository1.save(blog);
-       //userRepository1.save(user);
 
 
+        //Updating the userInformation and changing its blogs
+        List<Blog> currentBlogs = user.getBlogList();
+        currentBlogs.add(blog);
+        user.setBlogList(currentBlogs);
+
+        //Only calling the parent userRepository function as the child function will automatically be called by cascading
+        userRepository1.save(user);
     }
 
     public Blog findBlogById(int blogId){
@@ -60,24 +64,22 @@ public class BlogService {
     }
 
     public void addImage(Integer blogId, String description, String dimensions){
-        //add an image to the blog after creating it
-
+        //add an image to the blog
         Blog blog = blogRepository1.findById(blogId).get();
-
-        Image image = imageService1.createAndReturn(blog,description,dimensions);
-        image.setBlog(blog);
+        Image image = imageService1.createAndReturn(blog, description, dimensions);
         List<Image> imageList = blog.getImageList();
-
-        if(imageList == null)
-            imageList = new ArrayList<>();
         imageList.add(image);
         blog.setImageList(imageList);
+
         blogRepository1.save(blog);
+        //Just calling the parent repository and child repository will automatically be called.
 
     }
 
     public void deleteBlog(int blogId){
         //delete blog and corresponding images
-        blogRepository1.deleteById(blogId);
+        Blog blog = blogRepository1.findById(blogId).get();
+        if(blog != null)
+         blogRepository1.deleteById(blogId);
     }
 }
